+++ packages/file/lib/src/backends/memory/memory_file_system.dart
@@ -200,7 +200,7 @@ class _MemoryFileSystem extends FileSystem
   /// Gets the node backing for the current working directory. Note that this
   /// can return null if the directory has been deleted or moved from under our
   /// feet.
-  DirectoryNode get _current => findNode(cwd) as DirectoryNode;
+  DirectoryNode? get _current => findNode(cwd) as DirectoryNode?;
 
   @override
   Node? findNode(
@@ -211,7 +211,9 @@ class _MemoryFileSystem extends FileSystem
     List<String>? pathWithSymlinks,
     bool followTailLink = false,
   }) {
-    if (_context.isAbsolute(path)) {
+    if (path.isEmpty) {
+      return null;
+    } else if (_context.isAbsolute(path)) {
       reference = _root;
       path = path.substring(style.drive.length);
     } else {